Run automated garbage collections in foreground
git will automatically run an automated garbage collection in the background
after fetching, but it is not guaranteed that it will finish before the cache
is uploaded at the end of the workflow run.

Configure `gc.autoDetach=false` to force it run in the foreground. This will make
a single run with collection slower, but the cache will be (much) smaller.