Add a basic index
This is the original one slightly reworked so that we keep only one
version per branch. The panels were shortened due to less data being
displayed so there are now 3 per line. The status of each branch is
now reported (DEV/LTS/EOL), a neutral background is used for EOL, a
red one for development.

The links to dconv at the bottom were completed with the original
description and links arranged to be more visible.